refactor: Update frontmatter.json and time-line.ts
This commit updates the frontmatter.json file and the time-line.ts file. The changes in frontmatter.json include updating the title and meta description, as well as adding tags and keywords. In the time-line.ts file, a new type TFrontMatter is added to define the structure of the front matter data. These changes are made to improve the metadata and organization of the project.